requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
Django==3.2.13
django-extensions==3.1.5
django-livereload-server==0.4
djangorestframework==3.13.1
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.1
typing-extensions==4.1.1
